==> notebooks/1. translate_json_to_csv.ipynb <==
import json
import pandas

Extracts contents of poisk_in_ua_messages.json

Saves contents of messages dictionary to poisk_in_ua_messages.csv

with open("../poisk_in_ua/poisk_in_ua_messages.json", encoding="utf-8-sig") as file:
    dataset = json.load(file)["messages"]
    dataset_df = pandas.DataFrame(dataset)


    dataset_df.to_csv("../poisk_in_ua/poisk_in_ua_messages.csv")
