==> notebooks/4. translate_geocoded_dataset_to_geojson.ipynb <==
Load dataset

Load older datase

Merge datasets together

Output merged dataset to geojson

import pandas as pd, geopandas as gpd
from shapely.geometry import Point
import time
import calendar

timestamp = calendar.timegm(time.gmtime())
with open("../geolocated_datasets/1686573190-cleaned-of-birthdates-@poisk_in_ua.json", encoding="utf-8-sig") as raw_geodata_1:
    gdf_1 = gpd.read_file(raw_geodata_1)
gdf_2 = gpd.read_file("../geolocated_datasets/1686573840-geolocated_update.csv", encoding="utf-8-sig", crs="EPSG:3857")

gdf_2 = gdf_2[gdf_2['coordinates'].str.strip().astype(bool)]
gdf_2[["_latitude", "_longitude"]] = gdf_2["coordinates"].str.split(", ", expand = True)
gdf_2["coordinates"] = gdf_2["coordinates"].apply(lambda x: Point(x.split(', ')[1], x.split(", ")[0]))

gdf_2 = gdf_2.set_geometry("coordinates")
gdf_2.drop("geometry" , axis='columns', inplace = True)
gdf_2 = gdf_2.rename(columns={"coordinates":"geometry"})

gdf_1['id']=gdf_1['id'].astype(int)
gdf_2['id']=gdf_2['id'].astype(int)

gdf_2

merged_gdf = pd.concat([gdf_1, gdf_2], axis=0)

merged_gdf.drop(["Unnamed: 0", ], axis="columns", inplace=True)

merged_gdf

def _retrieve_missing_message(source_df, geolocated_df):
    source_df = pd.read_csv(source_df)
    source_df["id"] = source_df["id"].astype(int)
    source_df = source_df[['id', 'message']]
    temp = geolocated_df.merge(source_df, on="id", how='left')
    temp["message_x"].fillna("", inplace = True)
    temp["message_y"].fillna("", inplace = True)
    temp["message"] = temp["message_x"].map(str) + " " + temp["message_y"]
    temp.drop(columns=["message_x", "message_y"], axis=0, inplace = True)
    return temp

merged_gdf_1  = _retrieve_missing_message("../poisk_in_ua/poisk_in_ua_messages.csv", merged_gdf)

output_gdf = gpd.GeoDataFrame(merged_gdf_1)
output_gdf["date"] = output_gdf['date'].astype(str)

output_gdf.set_geometry(col='geometry', inplace=True) 

output_gdf_filename = f"../geolocated_datasets/{str(timestamp)}-geolocated-@poisk_in_ua.json"
output_file = open(output_gdf_filename, "w", encoding="utf-8-sig")
contents = output_gdf.to_json()
output_file.write(contents)
