==> notebooks/5. Remove located by birthdate.ipynb <==
import pandas as pd, geopandas as gpd
import calendar, time
import re
import json

In the first installment of this scripts, the system failed to geolocate some of messages due to incorrect parsing

In some instances, double posting or posting with the suffixes "g.r" or "g" was interpreted as a city, whereas these are abbreviations to "birthdate"

These ended up being erroneously geocoded by Nominatim

We'll first remove these results from the output dataset

Add a fail-safe to prevent geocoding birthdates instead of places

Re-geocode the failed geocoded messages by keeping their ids and re-parsing them


with open("../geolocated_datasets/1686571967-geolocated-@poisk_in_ua.json", encoding="utf-8-sig") as filec:
    original_gdf = gpd.read_file(filec, driver="GeoJSON")
timestamp = calendar.timegm(time.gmtime())

located_cities = original_gdf.loc[original_gdf["tag"]=="city"]

def find_dates(source):
    return re.search("\d{2}\.\d{2}\.\d{4}", source)


_test = located_cities
_test["test"] = located_cities["location"].apply(lambda a: find_dates(a))

identified_failures = _test.loc[_test["test"].notnull()]
identified_failures = identified_failures.drop("test", axis="columns")
identified_failures_1 = _test.loc[_test["message"].str.contains("Не идентифицирован", case = False, na = False)]

original_gdf["id"] = original_gdf["id"].astype(int)
identified_failures["id"] = identified_failures["id"].astype(int)
original_gdf["date"] = original_gdf["date"].astype(str)
identified_failures["date"] = identified_failures["date"].astype(str)
identified_failures_1["id"] = original_gdf["id"].astype(int)
identified_failures_1["date"] = identified_failures["date"].astype(str)

cleaned = original_gdf.drop(identified_failures.index, axis=0)
cleaned = test.drop(identified_failures_1.index, axis=0)

output_gdf_filename = f"../geolocated_datasets/{str(timestamp)}-cleaned-of-birthdates-@poisk_in_ua.json"
output_file = open(output_gdf_filename, "w", encoding="utf-8-sig")
contents = cleaned.to_json()
output_file.write(contents)

identified_failures

identified_failures_name = f"../to_geocode_datasets/{timestamp}-to_reparse.json"
output_file_1 = open(identified_failures_name, "w", encoding="utf-8-sig")
contents = identified_failures.to_json()
output_file_1.write(contents)
