==> notebooks/8. fatalities_reported_around_Russian_cities.ipynb <==
import pandas as pd, geopandas as gpd
import calendar, time
import json

Load fatalities dataset
Load cities shapefile

cities = gpd.read_file("../shp/Russian cities.shp",crs={'init' :'epsg:4326'})
cities = cities[["NAME", "geometry"]]
source = "../geolocated_datasets/1686574174R-geolocated-@poisk_in_ua.json"
with open(source, encoding = "utf-8-sig") as source_file:
    source_gdf = gpd.read_file(source_file, driver = "GeoJSON")

Convert both datasest to ESPG:3857 to allow use of meter system

source_gdf = source_gdf.to_crs(3857)

Create buffer of 10kms

buffer_distance = 10 * 1000
cities['lon'] = cities["geometry"].x
cities['lat'] = cities["geometry"].y
cities_buffer = cities.to_crs(3857)

cities_buffer.geometry = cities_buffer.buffer(buffer_distance)

Spatial join on buffers overlapping with fatalities dataset

pointsOnCity = gpd.sjoin(source_gdf, cities_buffer, how="right")
pointsOnCity['count']=1
fatalities_per_city = pointsOnCity.groupby('NAME').agg({'geometry':'first', 'lat':'first', 'lon':'first', 'count':'sum'}).reset_index()
fatalities_per_city.set_geometry("geometry", crs = "EPSG:4326", inplace = True)

fatalities_per_city

Export dataset to geolocated_dataset folder

timestamp = calendar.timegm(time.gmtime())
fatalities_per_city_fn = f"../geolocated_datasets/{timestamp}-geolocated_fatalities_per_city.json"
output_file = open(fatalities_per_city_fn, "w", encoding="utf-8-sig")
contents = fatalities_per_city.to_json()
output_file.write(contents)
