==> notebooks/9. statistics per region.ipynb <==
import pandas as pd, geopandas as gpd
import calendar, time


gd_file = "../geolocated_datasets/1687355357-fatalities_per_region_Russia.json"
gd_crimea = "../geolocated_datasets/1687355373-Crimea.json"
gd_donbas = "../geolocated_datasets/1687355377-Donbas.json"
with open(gd_file, encoding="utf-8-sig") as gd_contents:
    gdf = gpd.read_file(gd_contents, driver = "GeoJSON")

with open(gd_crimea, encoding = "utf-8-sig") as contents_1:
    gdf_crimea = gpd.read_file(contents_1, driver = "GeoJSON")

with open(gd_donbas, encoding = "utf-8-sig") as contents_2:
    gdf_donbas = gpd.read_file(contents_2, driver = "GeoJSON")

df = gdf[["name_en", "count"]]
df_crimea = gdf_crimea[["name_en", "count"]]
df_donbas = gdf_donbas[["name_en", "count"]]

timestamp = calendar.timegm(time.gmtime())
df.to_csv(f"../statistics/{timestamp}-statistics-per-Russian-region.csv")
df_crimea.to_csv(f"../statistics/{timestamp}-statistics-Crimea.csv")
df_donbas.to_csv(f"../statistics/{timestamp}-statistics-Donbas.csv")
